requirements.txt
playwright
pytest
pytest-html
requests
python-dotenv